+++ src/ChannelMap.php
@@ -22,7 +22,7 @@ trait ChannelMap
      */
     private function registerChannels()
     {
-        if ($this->app['config']->get('messenger.routing.channels.enabled')) {
+        if (config('messenger.routing.channels.enabled')) {
             $this->app->make(BroadcastManager::class)->routes($this->channelRouteConfiguration());
 
             $broadcaster = $this->app->make(Broadcaster::class);
@@ -41,9 +41,9 @@ trait ChannelMap
     private function channelRouteConfiguration(): array
     {
         return [
-            'domain' => $this->app['config']->get('messenger.routing.channels.domain'),
-            'prefix' => $this->app['config']->get('messenger.routing.channels.prefix'),
-            'middleware' => $this->mergeApiMiddleware($this->app['config']->get('messenger.routing.channels.middleware')),
+            'domain' => config('messenger.routing.channels.domain'),
+            'prefix' => config('messenger.routing.channels.prefix'),
+            'middleware' => $this->mergeApiMiddleware(config('messenger.routing.channels.middleware')),
         ];
     }
 }