+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 #### Added
 - bool column `embeds` and nullable/text column `extra` on `messages` table.
 - Optional extra payload when sending messages to allow custom json to be stored with the message.
+- More extensive model factories.
 
 #### Changed
 - Renamed action class UpdateMessage to EditMessage.